requirements.txt
from_root
python-dotenv
langchain_groq
langchain_openai